[refactor] member도메인 dto 리팩토링
문제 사항 :
- member도메인의 dto 전체적인 리팩토링
변경 사항:
- 필요없는 어노테이션, 메서드 삭제
- CreateMemberDTO에 검증 추가
