+++ src/main/java/com/fasttime/domain/member/dto/request/EditRequest.java
@@ -1,22 +1,18 @@
 package com.fasttime.domain.member.dto.request;
 
-import lombok.Data;
+import jakarta.validation.constraints.NotNull;
+import lombok.Getter;
 import lombok.NoArgsConstructor;
+import lombok.Setter;
 
-@Data
+@Getter
+@Setter
 @NoArgsConstructor
 public class EditRequest {
 
+    @NotNull
     private String nickname;
     private String image;
-
-    public String getImage() {
-        return image;
-    }
-
-    public void setImage(String image) {
-        this.image = image;
-    }
 
 
 }
