+++ src/main/java/com/fasttime/domain/member/service/MemberService.java
@@ -1,6 +1,6 @@
 package com.fasttime.domain.member.service;
 
-import com.fasttime.domain.member.dto.request.MyPageInfoDTO;
+import com.fasttime.domain.member.dto.response.MyPageInfoDTO;
 import com.fasttime.domain.member.dto.request.EditRequest;
 import com.fasttime.domain.member.dto.request.LoginRequestDTO;
 import com.fasttime.domain.member.exception.EmailAlreadyExistsException;
