+++ src/test/java/com/fasttime/domain/member/docs/MemberControllerDocsTest.java
@@ -27,7 +27,7 @@ import com.fasttime.docs.RestDocsSupport;
 import com.fasttime.domain.member.controller.MemberController;
 import com.fasttime.domain.member.dto.request.CreateMemberDTO;
 import com.fasttime.domain.member.dto.request.LoginRequestDTO;
-import com.fasttime.domain.member.dto.request.MyPageInfoDTO;
+import com.fasttime.domain.member.dto.response.MyPageInfoDTO;
 import com.fasttime.domain.member.entity.Member;
 import com.fasttime.domain.member.repository.MemberRepository;
 import com.fasttime.domain.member.dto.request.EditRequest;
